fix: css

+++ src/utils/locastorage.ts
@@ -1,12 +1,10 @@
 export const localStorage = {
     set: <T>(id: string, value: T) => {
         const stringifyValue = JSON.stringify(value)
-        if(typeof window !== "undefined")
-            window.localStorage.setItem(id, stringifyValue)
+        if (typeof window !== 'undefined') window.localStorage.setItem(id, stringifyValue)
     },
     get: <T>(id: string): T | null => {
-        if(typeof window !== "undefined")
-        {
+        if (typeof window !== 'undefined') {
             const value = window.localStorage.getItem(id)
             if (value) {
                 return JSON.parse(value) as T
@@ -14,11 +12,8 @@ export const localStorage = {
             return null
         }
         return null
-        
     },
-    reset: () => 
-    {
-        if(typeof window !== "undefined")
-            window.localStorage.clear()
+    reset: () => {
+        if (typeof window !== 'undefined') window.localStorage.clear()
     },
 }
